feat: 테마 생성, 제거를 어드민 기능으로 분리
