+++ src/main/java/nextstep/theme/AdminThemeController.java
@@ -1,0 +1,31 @@
+package nextstep.theme;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+
+@RestController
+@RequestMapping("/admin/themes")
+public class AdminThemeController {
+    private ThemeService themeService;
+
+    public AdminThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> createTheme(@RequestBody ThemeRequest themeRequest) {
+        Long id = themeService.create(themeRequest);
+        return ResponseEntity.created(URI.create("/themes/" + id))
+                .build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
+        themeService.delete(id);
+
+        return ResponseEntity.noContent()
+                .build();
+    }
+}
